--- index.js
@@ -1,17 +1,0 @@
-'use strict';
-
-let bubbleSort = require('./bubble-sort');
-let coverByTile = require('./entry-test/coverByTile');
-let UnionFind = require('./unionFind');
-let bitPlusPlus = require('./bitPlusPlus');
-let LinkedList = require('./linkedList');
-
-let algolib = {
-    bubbleSort,
-    coverByTile,
-    UnionFind,
-    bitPlusPlus,
-    LinkedList
-};
-
-module.exports = algolib;
