+++ methods/bitPlusPlus/bitPlusPlus.spec.js
