+++ methods/coverByTile/coverByTile.js
