+++ methods/coverByTile/tile.spec.js
