+++ methods/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+let bubbleSort = require('./bubble-sort/bubble-sort');
+let coverByTile = require('./coverByTile/coverByTile');
+let UnionFind = require('./unionFind/unionFind');
+let bitPlusPlus = require('./bitPlusPlus/bitPlusPlus');
+let LinkedList = require('./linkedList/linkedList');
+
+let algolib = {
+    bubbleSort,
+    coverByTile,
+    UnionFind,
+    bitPlusPlus,
+    LinkedList
+};
+
+module.exports = algolib;
